# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AskHealth AI Community Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Community {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconEmoji?: string | null;
  /** @nullable */
  coverColor?: string | null;
  isArchived: boolean;
  memberCount: number;
  postCount: number;
  createdAt: string;
}

export interface CreateCommunityBody {
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconEmoji?: string | null;
  /** @nullable */
  coverColor?: string | null;
}

export interface UpdateCommunityBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconEmoji?: string | null;
  /** @nullable */
  coverColor?: string | null;
  /** @nullable */
  isArchived?: boolean | null;
}

export interface CommunityStats {
  communityId: number;
  memberCount: number;
  postCount: number;
  commentCount: number;
  weeklyPosts: number;
  weeklyComments: number;
  weeklyNewMembers: number;
}

export interface Post {
  id: number;
  communityId: number;
  authorId: string;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  authorLevel?: number;
  title: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  isPinned: boolean;
  upvoteCount: number;
  commentCount: number;
  hasUpvoted: boolean;
  isBroadcast: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePostBody {
  title: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface UpdatePostBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface PinPostBody {
  isPinned: boolean;
}

export interface UpvoteResult {
  upvoteCount: number;
  hasUpvoted: boolean;
}

export interface Comment {
  id: number;
  postId: number;
  authorId: string;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  authorLevel: number;
  content: string;
  createdAt: string;
}

export interface CreateCommentBody {
  content: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  admin: "admin",
  moderator: "moderator",
  member: "member",
} as const;

export interface UserProfile {
  id: string;
  clerkId: string;
  displayName: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserProfileRole;
  isBanned: boolean;
  healthCredits: number;
  level: number;
  weeklyCredits: number;
  createdAt: string;
}

export interface UpdateUserBody {
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export type UpdateUserRoleBodyRole =
  (typeof UpdateUserRoleBodyRole)[keyof typeof UpdateUserRoleBodyRole];

export const UpdateUserRoleBodyRole = {
  admin: "admin",
  moderator: "moderator",
  member: "member",
} as const;

export interface UpdateUserRoleBody {
  role: UpdateUserRoleBodyRole;
}

export interface BanUserBody {
  isBanned: boolean;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  level: number;
  credits: number;
  badgeCount: number;
}

export interface Achievement {
  id: number;
  userId: string;
  badgeName: string;
  badgeDescription: string;
  badgeIcon: string;
  earnedAt: string;
}

export interface CreditsSummary {
  userId: string;
  healthCredits: number;
  weeklyCredits: number;
  level: number;
  levelName: string;
  creditsToNextLevel: number;
  progressPercent: number;
  recentBadges: Achievement[];
}

export interface SearchResults {
  posts: Post[];
  members: UserProfile[];
  totalCount: number;
}

export interface BroadcastBody {
  title: string;
  content: string;
  /** @nullable */
  communityIds?: number[] | null;
  postToAll: boolean;
}

export interface BroadcastResult {
  postsCreated: number;
  communityIds: number[];
}

export interface AdminStats {
  totalUsers: number;
  totalCommunities: number;
  totalPosts: number;
  totalComments: number;
  weeklyActiveUsers: number;
  newUsersThisWeek: number;
  topCommunities: Community[];
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface YuktiChatBody {
  message: string;
  /** @nullable */
  communityId?: number | null;
  /** @nullable */
  communityName?: string | null;
  history?: ChatMessage[];
}

export interface YuktiChatResponse {
  reply: string;
  /** @nullable */
  communityContext?: string | null;
}

export type ListCommunitiesParams = {
  includeArchived?: boolean;
};

export type ListPostsParams = {
  page?: number;
  limit?: number;
};

export type GetLeaderboardParams = {
  period?: GetLeaderboardPeriod;
};

export type GetLeaderboardPeriod =
  (typeof GetLeaderboardPeriod)[keyof typeof GetLeaderboardPeriod];

export const GetLeaderboardPeriod = {
  weekly: "weekly",
  all_time: "all_time",
} as const;

export type GlobalSearchParams = {
  q: string;
  type?: GlobalSearchType;
};

export type GlobalSearchType =
  (typeof GlobalSearchType)[keyof typeof GlobalSearchType];

export const GlobalSearchType = {
  all: "all",
  posts: "posts",
  members: "members",
} as const;