# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AskHealth AI Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /communities:
    get:
      operationId: listCommunities
      tags: [communities]
      summary: List all communities
      parameters:
        - name: includeArchived
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of communities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Community"
    post:
      operationId: createCommunity
      tags: [communities]
      summary: Create a new community (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommunityBody"
      responses:
        "201":
          description: Created community
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Community"
  /communities/{communityId}:
    get:
      operationId: getCommunity
      tags: [communities]
      summary: Get a community by ID
      parameters:
        - name: communityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Community
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Community"
        "404":
          description: Not found
    patch:
      operationId: updateCommunity
      tags: [communities]
      summary: Update a community (admin only)
      parameters:
        - name: communityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommunityBody"
      responses:
        "200":
          description: Updated community
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Community"
    delete:
      operationId: archiveCommunity
      tags: [communities]
      summary: Archive a community (admin only)
      parameters:
        - name: communityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived
  /communities/{communityId}/stats:
    get:
      operationId: getCommunityStats
      tags: [communities]
      summary: Get stats for a community (member count, post count, weekly activity)
      parameters:
        - name: communityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Community statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityStats"
  /communities/{communityId}/members:
    get:
      operationId: getCommunityMembers
      tags: [communities]
      summary: Get members of a community
      parameters:
        - name: communityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"

  # ─── Posts ──────────────────────────────────────────────────────────────
...
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ─── Users ───────────────────────────────────────────────────────────────
...
  /admin/broadcast:
    post:
      operationId: broadcastAnnouncement
      tags: [admin]
      summary: Broadcast an announcement to all or selected communities (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastBody"
      responses:
        "201":
          description: Broadcast created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastResult"
...
  /ai/chat:
    post:
      operationId: yuktiChat
      tags: [ai]
      summary: Chat with Yukti AI (community-aware health-ops assistant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YuktiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YuktiChatResponse"
components:
  schemas:
...
    Community:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        iconEmoji:
          type: ["string", "null"]
        coverColor:
          type: ["string", "null"]
        isArchived:
          type: boolean
        memberCount:
          type: integer
        postCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - isArchived
        - memberCount
        - postCount
        - createdAt
    CreateCommunityBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        iconEmoji:
          type: ["string", "null"]
        coverColor:
          type: ["string", "null"]
      required:
        - name
        - slug
    UpdateCommunityBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        iconEmoji:
          type: ["string", "null"]
        coverColor:
          type: ["string", "null"]
        isArchived:
          type: ["boolean", "null"]
    CommunityStats:
      type: object
      properties:
        communityId:
          type: integer
        memberCount:
          type: integer
        postCount:
          type: integer
        commentCount:
          type: integer
        weeklyPosts:
          type: integer
        weeklyComments:
          type: integer
        weeklyNewMembers:
          type: integer
      required:
        - communityId
        - memberCount
        - postCount
        - commentCount
        - weeklyPosts
        - weeklyComments
        - weeklyNewMembers
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        authorLevel:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - authorId
        - authorName
        - authorLevel
        - content
        - createdAt
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        displayName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, moderator, member]
        isBanned:
          type: boolean
        healthCredits:
          type: integer
        level:
          type: integer
        weeklyCredits:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - displayName
        - email
        - role
        - isBanned
        - healthCredits
        - level
        - weeklyCredits
        - createdAt
...
    BroadcastBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        communityIds:
          type: ["array", "null"]
          items:
            type: integer
        postToAll:
          type: boolean
      required:
        - title
        - content
        - postToAll
    BroadcastResult:
      type: object
      properties:
        postsCreated:
          type: integer
        communityIds:
          type: array
          items:
            type: integer
      required:
        - postsCreated
        - communityIds
...
    YuktiChatBody:
      type: object
      properties:
        message:
          type: string
        communityId:
          type: ["integer", "null"]
        communityName:
          type: ["string", "null"]
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required:
        - message
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content
    YuktiChatResponse:
      type: object
      properties:
        reply:
          type: string
        communityContext:
          type: ["string", "null"]
      required:
        - reply